delete user's QuestionBase

+++ src/service/QuestionBaseService.java
@@ -3,18 +3,18 @@ package service;
 import java.sql.ResultSet;
 import java.util.List;
 
-
 import java.util.ArrayList;
 import database.Connect;
 import domain.QuestionBase;
 
 public class QuestionBaseService {
-	private Connect cont = new Connect();
+	private Connect cont;
 	private List<QuestionBase> questionBases;
 	private QuestionBase questionBase;
 	
 	public QuestionBase getQuestionBase(int id){
 		String sql = "select * from questionbase where id=" + id;
+		cont = new Connect();
 		ResultSet result = cont.executeQuery(sql);	
 		
 		try{
@@ -31,28 +31,36 @@ public class QuestionBaseService {
 		return questionBase;
 	}
 	public int addQuestionBase(QuestionBase qBase) {
+		cont = new Connect();
 		String sql = "insert into questionbase(id, userID, title, description) values(" + qBase.getId() + ", " + qBase.getUserID() + ", '"
 				+ qBase.getTitle() + "', '" + qBase.getDescription() + "')";
 		
 		int id = cont.executeUpdateID(sql);
 		System.out.println("addQuestionBase sql: "+ sql + "   *id:" + id);
+		int in = 0;
 		if (id > 0) {
-			sql = "insert into user_questionbase(userId, questionBaseID) values(" + qBase.getUserID() + ", " + qBase.getId() + ")"; 
-			cont.executeUpdate(sql);
+			qBase.setId(id);
+			in = addUser_QuestionBase(qBase.getUserID(), qBase.getId());			
 		}
 		//System.out.println("LAST_INSERT_ID: " + id);
-		return id;
+		return in;
 	}
+
 	public List<QuestionBase> getUserQuestionBases(int userID) {
-		String sql = "select questionBaseID from user_questionbase where userID='" + userID + "'";
+		cont = new Connect();
+		String sql = "select questionBaseID from user_questionbase where userID=" + userID;
+		System.out.println("user_questionbase sql: " + sql);
 		ResultSet result = cont.executeQuery(sql);	
 		questionBases = new ArrayList<>();
 		try{
-			if (result.next()){
+			while (result.next()){
+				questionBase = new QuestionBase();
 				questionBase = getQuestionBase(result.getInt("questionBaseID"));
 				questionBases.add(questionBase);
+				System.out.println("get one qBase");
 			}
 		}catch (Exception e) {
+			e.printStackTrace();
 			questionBases = null;
 		}
 		return questionBases;
@@ -60,10 +68,27 @@ public class QuestionBaseService {
 	public int delQuestionBase(QuestionBase qBase) {
 		int userID = qBase.getUserID();
 		int id = qBase.getId();
-		
-		String sql = "delete from questionbase where id=" + id;
+		int i = delUser_QuestionBase(userID, id);
+		int j = -1;
+		if (i >= 0) {
+			cont = new Connect();
+			String sql = "delete from questionbase where id=" + id;
+			j = cont.executeUpdate(sql);
+		}
+		return j;
+	}
+	
+	public int addUser_QuestionBase(int userID, int qBaseID){
+		cont = new Connect();
+		String sql = "insert into user_questionbase(userID, questionBaseID) values(" + userID + ", " + qBaseID + ")"; 
 		int i = cont.executeUpdate(sql);
-		sql = "delete from user_questionbase where userID=" + userID + ", id=" + id;
+		System.out.println("insert user_questionbase sql: " + sql + " i: " + i);
+		return i;
+	}
+	public int delUser_QuestionBase(int userID, int qBaseID){
+		cont = new Connect();
+		String sql = "delete from user_questionbase where userID=" + userID + " and questionBaseID=" + qBaseID;
+		int i = cont.executeUpdate(sql);		
 		return i;
 	}
 }
